Refactor validators to inherit validation of IFormFile
(cherry picked from commit d328d8543db9bab96b515c71522b2023d5ba219c)
(cherry picked from commit ff5bcf2e007f2a2ddde6f58dde38da9bc94b58e6)
