+++ src/Equinor.Procosys.Preservation.WebApi/Controllers/Tags/UploadBaseDto.cs
@@ -1,0 +1,9 @@
+﻿using Microsoft.AspNetCore.Http;
+
+namespace Equinor.Procosys.Preservation.WebApi.Controllers.Tags
+{
+    public abstract class UploadBaseDto
+    {
+        public IFormFile File { get; set; }
+    }
+}
