+++ src/Equinor.Procosys.Preservation.WebApi/Controllers/Tags/UploadWithOverwriteAttachmentDto.cs
@@ -1,0 +1,7 @@
+﻿namespace Equinor.Procosys.Preservation.WebApi.Controllers.Tags
+{
+    public class UploadWithOverwriteAttachmentDto : UploadAttachmentDto
+    {
+        public bool OverwriteIfExists { get; set; }
+    }
+}
